Refactor fswatcher into own watcher
