+++ .gitignore
@@ -7,6 +7,7 @@
 
 # Ignore built artifact
 fits
+
 # Test binary, built with `go test -c`
 *.test
