+++ watchers/FsWatcher.go
@@ -1,0 +1,55 @@
+package watchers
+
+import (
+	"fmt"
+	"io"
+	"os"
+)
+
+type FsWatcher struct {
+	From string
+	To string
+}
+
+func MoveFile(from, to string) error {
+	in, err := os.Open(from)
+	if err != nil {
+		return fmt.Errorf("Couldn't open source file: %s", err)
+	}
+
+	out, err := os.Create(to)
+	if err != nil {
+		in.Close()
+		return fmt.Errorf("Couldn't open dest file: %s", err)
+	}
+
+	defer out.Close()
+
+	_, err = io.Copy(out, in)
+	in.Close()
+	if err != nil {
+		return fmt.Errorf("Writing to output file failed: %s", err)
+	}
+
+	err = out.Sync()
+	if err != nil {
+		return fmt.Errorf("Sync error: %s", err)
+	}
+
+	si, err := os.Stat(from)
+	if err != nil {
+		return fmt.Errorf("Stat error: %s", err)
+	}
+
+	err = os.Chmod(to, si.Mode())
+	if err != nil {
+		return fmt.Errorf("Chmod error: %s", err)
+	}
+
+	err = os.Remove(from)
+	if err != nil {
+		return fmt.Errorf("Failed removing original file: %s", err)
+	}
+
+	return nil
+}
